implement opponent detection and opponent controles
