+++ client/src/GameData/initGameState.js
@@ -1,12 +1,12 @@
-export default function initGameState() {
+export default function initGameState(state) {
   const game = {
     ball: {
-      x: 100,
-      y: 150,
+      x: 200,
+      y: 145,
       w: 10,
       h: 10,
-      dx: 0,
-      dy: 0,
+      dx: -1,
+      dy: -1,
       pdx: 2
     },
     player: {
@@ -20,18 +20,22 @@ export default function initGameState() {
     opponent: {
       x: 290 / 2 - 50,
       y: 40,
-      w: 50,
+      w: 100,
       h: 10,
       dx: 3,
       dy: 0
     },
     global: {
-      x: 295,
+      x: 300,
       y: 400,
       cheerWin: "You Won!!!",
       cheerLoss: "You Lost...",
       winner: "opponent"
     }
   };
+  if (state) {
+    game.ball.dy *= -1;
+    debugger;
+  }
   return game;
 }
