+++ lib/socket.js
@@ -1,4 +1,5 @@
 const socketIo = require("socket.io");
+const { gameCountDown } = require("./socket-callbacks");
 
 function initSocket(httpServer) {
   const io = socketIo(httpServer);
@@ -65,9 +66,9 @@ function initSocket(httpServer) {
             "new server message",
             sendServerChatMessage(allClients[socket.id], "join")
           );
-        socket.broadcast.to(room).emit("set player", "player 1");
-        io.in(room).emit("game ready", "Let users be able to click ready!");
-        console.log(sendServerChatMessage(allClients[socket.id], "join"));
+
+        socket.broadcast.to(room).emit("set opponent", "you are the opponent");
+        socket.emit("set player", "you are the player");
       } else {
         socket.emit("room full", "got to main");
         socket.leave(room);
@@ -94,23 +95,7 @@ function initSocket(httpServer) {
       console.log(message);
     });
 
-    function gameCountDown() {
-      let count = 4;
-      const timer = setInterval(() => {
-        io.in(room).emit("new server message", {
-          user: "socket.io - server",
-          nickname: "Server",
-          content: `${count - 1} seconds until start`
-        });
-        count -= 1;
-        if (count === 0) {
-          clearInterval(timer);
-          io.in(room).emit("game start", "start");
-        }
-      }, 1000);
-    }
-
-    socket.on("start game", gameCountDown);
+    socket.on("start game", () => gameCountDown(io, room));
     socket.on("conceded", () =>
       socket.broadcast.to(room).emit("opponent conceded", "opponent conceded")
     );
